Add npm package wrapper
This commit will introduce a wrapper around the C++ matching algorithm in form of an NPM package which is written in typescript and uses the Node Addon API to call the C++ matching algorithm from code.
This commit will add all required files, including the package.json and the binding.gyp file.