Add npm package tests
This will add tests to test the NPM package and whether it works and correctly interacts with the C++ matching algorithm.
Note: Because this tests the wrapper around the matching algorithm, it is possible that if the C++ code has errors, they can be detected using that tests as well!