Export types used for interacting with module
This will export the most important types used by the module internally and used by the module in the exported functions (the function that represent the externally available functionally of the module). This way, packages using this module can use those types in other variable declarations as part of their code.